chore: move 2fa to settings
