+++ packages/ps2/public/locales/en/settings.json
@@ -1,0 +1,28 @@
+{
+  "update-password": {
+    "title": "Update Password",
+    "description": "Update the password of your Zignaly Account.\nFor your security, you will be logged out after clicking Update password.",
+    "current-password": "Current Password:",
+    "new-password": "New Password:"
+  },
+  "enable-2fa": {
+    "title": "Two-Factor Authentication (2FA)",
+    "description": "Secure your account setting up two-factor authentication (2FA). Right now, 2FA is considered the best security practice, read more <0>here.</0>",
+    "setup-description": "Download the <0>Google Authentication App</0> and scan the QR code, or paste the Key Phrase manually into the app. For security you will be logged out after introducing the 6 digit code.",
+    "enter-code": "Enter the 6-digit code from the app:",
+    "enter-code-subtitle": "You'll need it every time you log in.",
+    "key-phrase": "Key Phrase:",
+    "key-phrase-copied": "Key Phrase copied to clipboard",
+    "key-phrase-info": "The Key Phrase is used to recover Google Authenticator in the event of a loss or change of device - please make sure to keep the key phrase safe before introducing the 6-digit code.",
+    "code-2fa": "2FA Code",
+    "setup-2fa": "Set up 2FA",
+    "enable-2fa": "Enable 2FA",
+    "success": "2 Factor Authentication enabled. Please log in again."
+  },
+  "disable-2fa": {
+    "title": "Disable 2FA",
+    "description": "Your account is secured with Two-Factor Authentication (2FA). To disable 2FA, enter the code from your authenticator app.",
+    "success": "2FA has been disabled successfully.",
+    "security": "For security, we recommend to have 2FA enabled in your account."
+  }
+}
