+++ packages/ps2/src/util/i18next.ts
@@ -61,6 +61,7 @@ i18n
       'withdraw',
       'withdraw-crypto',
       'withdraw-your-investment',
+      'settings',
     ],
     supportedLngs: supportedLanguages,
     defaultNS: 'common',
