+++ packages/ps2/src/views/Auth/components/LoginForm/index.tsx
@@ -7,16 +7,9 @@ import { LoginValidation } from './validations';
 import { useAuthenticate } from '../../../../apis/user/use';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { ROUTE_FORGOT_PASSWORD, ROUTE_SIGNUP } from '../../../../routes';
-import {
-  Button,
-  IconButton,
-  TextButton,
-  Typography,
-  ZigInput,
-} from '@zignaly-open/ui';
-import { Box, InputAdornment } from '@mui/material';
+import { Button, TextButton, Typography, ZigInput } from '@zignaly-open/ui';
+import { Box } from '@mui/material';
 import { LoginPayload } from '../../../../apis/user/types';
-import { Visibility, VisibilityOff } from '@mui/icons-material';
 import PasswordVisibilityAdornment from '../atoms/PasswordVisibilityAdornment';
 
 const LoginForm: React.FC = () => {
@@ -91,12 +84,10 @@ const LoginForm: React.FC = () => {
               type={showPassword ? 'text' : 'password'}
               InputProps={{
                 endAdornment: (
-                  <InputAdornment position='end'>
-                    <PasswordVisibilityAdornment
-                      show={showPassword}
-                      onToggle={() => setShowPassword(!showPassword)}
-                    />
-                  </InputAdornment>
+                  <PasswordVisibilityAdornment
+                    show={showPassword}
+                    onToggle={() => setShowPassword(!showPassword)}
+                  />
                 ),
               }}
               {...field}
